+++ lib/src/docker_commander_base.dart
@@ -6,7 +6,7 @@ import 'docker_commander_host.dart';
 /// The Docker manager.
 class DockerCommander extends DockerCMDExecutor {
   /// Current version of package `docker_commander`.
-  static final String VERSION = '2.0.9';
+  static final String VERSION = '2.0.10';
 
   /// Docker machine host.
   final DockerHost dockerHost;
@@ -363,8 +363,8 @@ class DockerCommander extends DockerCMDExecutor {
 
   /// Executes a formula function.
   Future<dynamic> formulaExec(String formulaName, String functionName,
-          [List? arguments]) =>
-      dockerHost.formulaExec(formulaName, functionName, arguments);
+          [List? arguments, Map<String, dynamic>? fields]) =>
+      dockerHost.formulaExec(formulaName, functionName, arguments, fields);
 }
 
 typedef DockerContainerInstantiator = DockerContainer? Function(